+++ GridKit/Model/PhasorDynamics/Load/Load.hpp
@@ -64,6 +64,7 @@ namespace GridKit
       virtual int allocate() override final;
       virtual int initialize() override final;
       virtual int tagDifferentiable() override final;
+      virtual int setAbsoluteTolerance(RealT) override final;
       virtual int evaluateResidual() override final;
       virtual int evaluateJacobian() override final;