+++ GridKit/Model/PhasorDynamics/SystemModel.hpp
@@ -49,10 +49,9 @@ namespace GridKit
       using PhasorDynamics::Component<scalar_type, index_type>::y_;
       using PhasorDynamics::Component<scalar_type, index_type>::yp_;
       using PhasorDynamics::Component<scalar_type, index_type>::tag_;
+      using PhasorDynamics::Component<scalar_type, index_type>::abs_tol_;
       using PhasorDynamics::Component<scalar_type, index_type>::f_;
       using PhasorDynamics::Component<scalar_type, index_type>::J_;
-      using PhasorDynamics::Component<scalar_type, index_type>::rel_tol_;
-      using PhasorDynamics::Component<scalar_type, index_type>::abs_tol_;
       using PhasorDynamics::Component<scalar_type, index_type>::variable_indices_;
       using PhasorDynamics::Component<scalar_type, index_type>::residual_indices_;