+++ GridKit/Model/PowerElectronics/CircuitComponent.hpp
@@ -259,17 +259,6 @@ namespace GridKit
       return size_opt_;
     }
 
-    void setTolerances(RealT& rel_tol, RealT& abs_tol) const final
-    {
-      rel_tol = rel_tol_;
-      abs_tol = abs_tol_;
-    }
-
-    void setMaxSteps(IdxT& msa) const final
-    {
-      msa = max_steps_;
-    }
-
     std::vector<ScalarT>& y() final
     {
       return y_;
@@ -298,6 +287,16 @@ namespace GridKit
     const std::vector<bool>& tag() const final
     {
       return tag_;
+    }
+
+    std::vector<ScalarT>& absoluteTolerance() final
+    {
+      return abs_tol_;
+    }
+
+    const std::vector<ScalarT>& absoluteTolerance() const final
+    {
+      return abs_tol_;
     }
 
     std::vector<ScalarT>& yB() final
@@ -437,6 +436,7 @@ namespace GridKit
     std::vector<ScalarT> y_;
     std::vector<ScalarT> yp_;
     std::vector<bool>    tag_;
+    std::vector<ScalarT> abs_tol_;
     std::vector<ScalarT> f_;
     std::vector<ScalarT> g_;
 
@@ -451,9 +451,6 @@ namespace GridKit
 
     RealT time_;
     RealT alpha_;
-
-    RealT rel_tol_;
-    RealT abs_tol_;
 
     IdxT max_steps_;