+++ GridKit/Model/PowerFlow/Branch/Branch.hpp
@@ -53,6 +53,7 @@ namespace GridKit
     int allocate();
     int initialize();
     int tagDifferentiable();
+    int setAbsoluteTolerance(RealT);
     int evaluateResidual();
     int evaluateJacobian();
     int evaluateIntegrand();