+++ GridKit/Model/PowerFlow/Bus/BusPQ.hpp
@@ -25,6 +25,7 @@ namespace GridKit
     using BaseBus<ScalarT, IdxT>::f_;
     using BaseBus<ScalarT, IdxT>::fB_;
     using BaseBus<ScalarT, IdxT>::tag_;
+    using BaseBus<ScalarT, IdxT>::abs_tol_;
 
   public:
     using RealT   = typename ModelEvaluatorImpl<ScalarT, IdxT>::RealT;
@@ -37,6 +38,7 @@ namespace GridKit
 
     virtual int allocate();
     virtual int tagDifferentiable();
+    virtual int setAbsoluteTolerance(RealT);
     virtual int initialize();
     virtual int evaluateResidual();
     virtual int initializeAdjoint();