+++ GridKit/Model/PowerFlow/MiniGrid/MiniGrid.cpp
@@ -31,8 +31,6 @@ namespace GridKit
       B23_(12.0)
   {
     // std::cout << "Create a load model with " << size_ << " variables ...\n";
-    rel_tol_ = 1e-5;
-    abs_tol_ = 1e-5;
   }
 
   template <class ScalarT, typename IdxT>