+++ GridKit/Model/PowerFlow/MiniGrid/MiniGrid.hpp
@@ -18,8 +18,6 @@ namespace GridKit
     using ModelEvaluatorImpl<ScalarT, IdxT>::time_;
     using ModelEvaluatorImpl<ScalarT, IdxT>::y_;
     using ModelEvaluatorImpl<ScalarT, IdxT>::f_;
-    using ModelEvaluatorImpl<ScalarT, IdxT>::rel_tol_;
-    using ModelEvaluatorImpl<ScalarT, IdxT>::abs_tol_;
 
     using RealT = typename ModelEvaluatorImpl<ScalarT, IdxT>::RealT;
 
@@ -31,6 +29,11 @@ namespace GridKit
     int initialize();
 
     int tagDifferentiable()
+    {
+      return -1;
+    }
+
+    int setAbsoluteTolerance(RealT)
     {
       return -1;
     }